dividend type should be unsigned in get_hash

+++ hashtable/hashtable.c
@@ -35,7 +35,7 @@ static int hash_function(const struct YASL_Object s, const int a, const int m) {
 static unsigned int get_hash(const struct YASL_Object s, const int num_buckets, const int attempt) {
 	const int hash_a = hash_function(s, PRIME_A, num_buckets);
 	if (attempt == 0) {
-		return hash_a % num_buckets;
+		return ((unsigned int)hash_a) % num_buckets;
 	}
 	const int hash_b = hash_function(s, PRIME_B, num_buckets);
 	return ((unsigned int) (hash_a + (attempt * (hash_b + (hash_b == 0))))) % num_buckets;
